DEV-27: Добавил инициализцию всех обработчиков (кроме finnish) переделал создание таблиц
