+++ data/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -8,4 +8,4 @@ class User(Base):
 
     id = Column(Integer, primary_key=True)
     username = Column(String)
-    chat_id = Column(Integer)
+    chat_id = Column(BigInteger)
